requirements.txt
hydra-core
omegaconf
numpy
torch
torchvision